Defensive timeouts against network partitions
Sometimes machines do not cleanly close connections (machine failure mostly)
or they are impossible to reach for other reasons. So we set TCP_USER_TIMEOUT
which prevents Cassandra from hanging and potentially going
unavailable under such conditions. This overrides the kernel level
retransmit policy to limit the amount of time that packets can remain
unacknowledged on a socket before giving up and closing that connection.

The default settings are 2s connection timeout and 30s unacknowledged
data timeout in Linux (the kernel closes the connection after ~50s of
unacknowledged data on a TCP connection). This is configurable both in
cassandra.yaml and via dynamic JMX endpoint on StorageService.