Repair Scheduler 3: Configuration
Introduces the two key pieces of configuration and context that are
needed by all future Repair Scheduler Classes. The first is the
configuration itself (e.g. Yaml config) and the second is a rolled up
Ineraction, Local Session and Config (which we call
RepairSchedulerContext) used by all the downstream classes.