Improve Keycache load speeds by a factor of the number of SSTables
Previously we were doing a linear traversal of the SSTables in a column
family for each key that we deserialized from the saved keycache. This
accidentally turned KeyCache loading into an O(n*m) algorithm instead of
and O(n) algorithm, which is particularly bad when the number of
sstables is high (for example LCS).

patch by Joseph Lynch